Separate out auth settings to allow for easily excluding config from git.
There are some secrets and per-environmnet settings for authorizing to
Salesforce that we don't want committed. We're going to gitignore those
but we also want to be able to migrate regular Salesforce settings, like
the API version, so we're separating out the per-environment auth
settings into a separate config.

I've also fixed up some broken config for the salesforce_pull module
along with these changes.