Another attempt at getting the config approach correct.
We're moving some config to state variables instead as they store
timestamps for last sync and delete info that we don't want to treat
like site config. We're also moving the salesforce_pull webhook key to
the salesforce.auth_settings config to keep it together with the rest of
the authentication credentials.